Quotes now support let-expression form.
In the process, I changed around the default quote usage.
Now, we should no longer be pasting arbitrary AST nodes into
liszt code as Lua objects.  The only valid AST node to paste in
is a QuoteExpr.  Eventually we may want to just wrap these in
some kind of explicit Quote object for consistency with other
inline-able Lua objects (e.g. macros, relations).

In keeping with the above, parameter ASTs passed to Macro generator
functions will also be wrapped in QuoteExpr AST nodes. (If you think
about this a bit, the idea of the change should be clearer.)