Added periodic boundary conditions, high-order scheme compatibility, calculation of deltaTime; corrected inviscid fluxes
The addition of periodic boundary conditions makes now the code compatible with
high-order spatial discretization. The xoffset*, yoffset* and signDouble
variables have been added, so that it is easier to switch back and forth between
periodic and symmetric boundary conditions. Particles are also periodic.
Note that to use periodic boundary conditions, the offsets are set equal to the
interior grid size in each direction. This is something to be discussed in
tomorrow's meeting, particularly how that affects the retrieval of the internal
'stencil' by Liszt, by relative references to needed cells. In the mean time,
it's been coded in this way to have the capability ready.

The computation of inviscid fluxes has been corrected to properly account the
fpdiag term, which was missing in earlier versions.

The deltaTime is now calculated at each time step based on the spectral radii
and the input cfl condition. Note that this deltaTime is for now a function of
the flow solver only, and the contribution from the particles should be added as
well.

A way to retrieve the maximum value of a field has been hard coded for now.
Ideally this will be taken out of this high-level application and brought as a
built-in function. In talking with Gilbert, it seems there are already plans for
doing so in future versions of Liszt.