Added particles feeder and collectors, reorganized code for better readibility, generalized initialization
Currently the total number of particles is defined at start-up. A state field
defines whether they are not fed yet (0), fed and active (1) or already
collected (2). For now, kernels related to particles include 'if' statements to
apply them only on the subset of particles that require it (for example, time
integration applies only to active (state == 1) particles. This could be improved
if/when subsets of a relation can be defined dynamically, based on the status of
a field defined on the particles.

Specialization of some kernels based on the options specified by the users is
done based on global vector variables deined as
L.NewGlobal(L.vector(L.double,n)), where n depends on the specialization (given
the number of parameters needed to define the case).

Global index variables (defined at the top of the file) are also used a la enum,
so that the specialization is done within a kernel through "if then" statements
that compare against those global variables. We discussed this in a prior
PSAAP2-CS meeting (whether to bring the specialization inside or outside of the
Liszt kernels, avoiding "if" branching inside them), and it seems that this
approach is correct. We will have to evaluate this when doing performance tests.

The code has been reorganized for better readibility, bringing all Liszt Kernels
together, split by objects (actually, namespaces for now) and the high-level
Lua-like functions in a separate part of the code, below the kernels).

Besides the TaylorGreen vortex flow initialization, the Uniform type has been
added, specifically for the UQCase that has already been implemented for 2D.
This case consists of a uniform background flow and two particle injectors that
require to specify a geometry (for now box-like), velocity of particles at
injection, and frequency of particle injection (with respect to the time steps).
The particles are randomly introduced within the injector geometry limits. The
initial distribution is split into two groups (a field named groupID is defined
to contain this information related to the injector that the particle came
from). The number of particles in each group is calculated based on the ratios
of injection frequencies, so that the injection lasts for approximately the same
time. This can be easily changed to the same number of particles for each
feeder/injector or any other option. The diamater of the particles is currently
initialized to a random distribution with a mean and a maximum deviation, but
can be easily modified to be injector-dependent.

A general feeder that initializes the particles to a random location within a
specified box also exists. Particles are not injected over time, but only at the
initial time step.

Regarding collectors, two types have been implemented:
CollectorNone means that there is no particle collection: it lets the particles
evolve without ever collecting them (which would imply setting their state to 2)
CollectorOutOfBox collects particles that exit the boundaries of a predefined
box geometry. The state of those particles is then set to 2, and they are not
active in the time integration (they are frozen at the position that they were
collected, without contributing to the flow solver, for example, if any two-way
coupling existed).

Additional fields are written as output for the particles containing the
'state', 'group' and 'id' fields. 'group' indicates, for the UQCase, the
injector that the particle comes from, whereas 'id' identifies the particle by
its indes (note that this is unique as the total number of particles is
initialized at once during the setup).

The UQCase