Added secondary string arguments to user functions; Closes #34
The design mostly conforms to the proposal.  No assignment of string
literals is allowed.  Special holes were provided for with Macros and
User functions (still always inlined) so that string literal arguments
can be propagated through.  String arguments must be untyped.
I also added the ability to write inline string values in Liszt code,
or inline by capturing a Lua upvalue, but did not add support for
format strings in print statements so #36 is still unresolved.

Finally, there is a new test for this functionality.